concord/src/requirements.txt
autobahn==0.10.5.post2
backports.ssl-match-hostname==3.4.0.2
cachetools==1.0.3
cffi==1.1.2
concord-py==0.2.15
cryptography==1.0
enum34==1.0.4
graphviz==0.4.3
idna==2.0
ipaddress==1.0.14
kafka-python==0.9.4
kazoo==2.0
pyasn1==0.1.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
six==1.9.0
thrift==0.9.2
trace2html==0.2.1
Twisted==15.3.0
txaio==1.0.3
urllib3==1.11
wheel==0.24.0
zope.interface==4.1.2
